+++ backend/IncomeCalculator-backend/src/main/java/pl/twyszomirski/service/ExchangeRateService.java
@@ -1,5 +1,6 @@
 package pl.twyszomirski.service;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -7,7 +8,7 @@ import java.util.Date;
  */
 public interface ExchangeRateService {
 
-    Float getExchangeRate(String fromCode, String ToCode, String dateIdentifier) throws NoExchangeRateException;
+    BigDecimal getExchangeRate(String fromCode, String ToCode, String dateIdentifier) throws NoExchangeRateException;
 
     String getDateToken(Date date);
 }
