+++ backend/IncomeCalculator-backend/src/main/java/pl/twyszomirski/service/YahooExchangeRateServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 import pl.twyszomirski.yahoo.finance.dto.ExchangeRateDto;
 
+import java.math.BigDecimal;
 import java.text.SimpleDateFormat;
 import java.util.Arrays;
 import java.util.Date;
@@ -31,7 +32,7 @@ public class YahooExchangeRateServiceImpl implements ExchangeRateService{
 
     @Override
     @Cacheable("exchangeRates")
-    public Float getExchangeRate(String fromCode, String toCode, String dateIdentifier) throws NoExchangeRateException{
+    public BigDecimal getExchangeRate(String fromCode, String toCode, String dateIdentifier) throws NoExchangeRateException{
 
         LOGGER.debug("Calling getExchangeRate with parameters {}, {}, {}", fromCode, toCode, dateIdentifier);
 
@@ -42,10 +43,10 @@ public class YahooExchangeRateServiceImpl implements ExchangeRateService{
 
         String url= EXCHANGE_RATE_SERVICE_URL.replaceAll(FROM_CURRENCY_PLACE_HOLDER, fromCode).replaceAll(TO_CURRENCY_PLACE_HOLDER, toCode);
 
-        Float result = null;
+        BigDecimal result = null;
         try {
             ResponseEntity<ExchangeRateDto> response = restTemplate.exchange(url, HttpMethod.GET, request, ExchangeRateDto.class);
-            result = Float.parseFloat(response.getBody().getQuery().getResults().getRate().getRate());
+            result = BigDecimal.valueOf(Float.parseFloat(response.getBody().getQuery().getResults().getRate().getRate()));
         }
         catch(Exception e){
             LOGGER.error("The damn yahoo finance service failed",e);
